risk-scorer-ml/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn==1.7.0
xgboost
joblib
pydantic
motor
python-dotenv